Move the doc site download to a submodule (#18)
Ali Ayas has already made a clean version, and we can guarantee consistent output by relying on certain versions of the site.

+++ Makefile
@@ -1,42 +1,27 @@
-source_link := https://www.sublimetext.com/docs/index.html
-local_path := www.sublimetext.com
+submodule_path := SublimeText_Documentation
+local_path := $(submodule_path)/www.sublimetext.com
 local_index := $(local_path)/docs/index.html
-built_path := $(local_path)/sublime-text.docset
+docset := sublime-text.docset
+dashing_json := $(local_path)/dashing.json
+built_path := $(local_path)/$(docset)
 
 .PHONY: all
-all: clean download pre-build build
-
-.PHONY: download
-download:
-	-wget \
-		--convert-links \
-		--recursive \
-		--page-requisites \
-		--no-parent \
-		--timestamping \
-		"$(source_link)"
+all: clean pre-build build
 
 .PHONY: fix
-pre-build: fix-html fix-css
+pre-build: fix-html
 
 .PHONY: fix-html
 fix-html:
 	python fix_html.py
 
-.PHONY: fix-css
-fix-css:
-	$(shell for f in $$(ls $(local_path)/*.*\?*); do mv "$$f" "$${f%\?*}"; done )
-	$(shell for f in $$(ls $(local_path)/**/*.*\?*); do mv "$$f" "$${f%\?*}"; done )
-
 build:
-	yq -j . dashing.yml > $(local_path)/dashing.json
+	yq -j . dashing.yml > $(dashing_json)
 	cd $(local_path) \
 	&& dashing build
 
 .PHONY: clean
 clean:
-	-rm -r $(local_path)
-
-.PHONY: clean-more
-clean-more: clean
-	-rm -r $(built_path)
+	[ -d "$(built_path)" ] && rm -r $(built_path) || true
+	[ -f "$(dashing_json)" ] && rm $(dashing_json) || true
+	git restore $(submodule_path) --recurse-submodules
